feat: integrate file hashing into Eleventy build process
- Add afterBuild hook in .eleventy.js for automatic file hashing
- Import hashing functions from hash-assets.js
- Remove separate hash-assets call from build script
- Remove unnecessary main() function from hash-assets.js
- Add cleanupOriginalFiles to afterBuild hook
- Hashing now runs automatically after Eleventy build in production mode
- Simplify build process by eliminating manual hashing step
- Development mode correctly skips hashing
